% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_plot.R
\name{anova_plot}
\alias{anova_plot}
\title{ANOVA Plot}
\usage{
anova_plot(model, predictor = NULL, graph_label_name = NULL)
}
\arguments{
\item{model}{fitted model (usually `lm` or `aov` object)}

\item{predictor}{predictor variable (must specified for non-interaction plot)}

\item{graph_label_name}{graph label name}
}
\value{
plot object
}
\description{
ANOVA Plot
}
\examples{
fit = iris \%>\% lm(data = ., Sepal.Length ~ Species)
anova_plot(fit,predictor = Species)

}
