% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-interaction_plot_check.R
\name{interaction_plot_check}
\alias{interaction_plot_check}
\title{interaction_plot_check}
\usage{
interaction_plot_check(interaction_term)
}
\arguments{
\item{interaction_term}{}
}
\value{
nothing
}
\description{
interaction_plot_check
}
\keyword{internal}
